+++ src/library/DIPS.Mobile.UI/API/Camera/BarcodeScanning/DidFindBarcodeCallback.cs
@@ -1,3 +1,0 @@
-namespace DIPS.Mobile.UI.API.Camera.BarcodeScanning;
-
-public delegate void DidFindBarcodeCallback(BarcodeScanResult barcodeScanResult);